Data and code to convert the file format
The data viewer runs off a SQLLite database, this allows the app to quickly return the correct table the user selects. This comes at a cost; the RDS file is 88MB, the SQLLite is ~7GB. 

The code will allow you to convert the RDS file into an SQLLite file to use with the app.